+++ upstairs/src/test.rs
@@ -1,5 +1,6 @@
 // Copyright 2023 Oxide Computer Company
 
+use crate::guest::Guest;
 use crate::*;
 
 /*
@@ -690,429 +691,6 @@ pub(crate) mod up_test {
 
         assert_eq!(successful_hash, Some(read_response_hash));
         assert_eq!(read_response.data, original_data);
-
-        Ok(())
-    }
-
-    async fn assert_consumed(io: &mut GuestIoHandle) {
-        tokio::select! {
-            _ = io.recv() => {
-                // correct!
-            },
-            _ = tokio::time::sleep(std::time::Duration::from_millis(1)) => {
-                panic!("timed out while waiting for message");
-            }
-        }
-    }
-
-    async fn assert_none_consumed(io: &mut GuestIoHandle) {
-        tokio::select! {
-            _ = io.recv() => {
-                panic!("got message when expecting nothing")
-            },
-            _ = tokio::time::sleep(std::time::Duration::from_millis(1)) => {
-                // nothing to do here
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_no_iop_limit() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-        assert_none_consumed(&mut io).await;
-
-        // Don't use guest.read, that will send a block size query that will
-        // never be answered.
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(1, 512),
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(8, 512),
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(32, 512),
-            })
-            .await;
-
-        // With no IOP limit, all requests are consumed immediately
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-
-        assert_none_consumed(&mut io).await;
-
-        // If no IOP limit set, don't track it
-        assert_eq!(io.iop_tokens, 0);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_iop_limit() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-        io.set_iop_limit(16000, 2);
-
-        assert_none_consumed(&mut io).await;
-
-        // Don't use guest.read, that will send a block size query that will
-        // never be answered.
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(1, 512),
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(8, 512),
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(31, 512),
-            })
-            .await;
-
-        // First two reads succeed
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-
-        // Next cannot be consumed until there's available IOP tokens so it
-        // remains in the queue.  Strictly speaking, it's been popped to the
-        // `req_head` position.
-        assert_none_consumed(&mut io).await;
-        assert!(io.req_rx.try_recv().is_err());
-        assert_eq!(io.iop_tokens, 2);
-        assert!(io.req_head.is_some());
-
-        // Replenish one token, meaning next read can be consumed
-        io.leak_iop_tokens(1);
-        assert_eq!(io.iop_tokens, 1);
-
-        assert_consumed(&mut io).await;
-        assert!(io.req_rx.try_recv().is_err());
-        assert!(io.req_head.is_none());
-        assert_eq!(io.iop_tokens, 2);
-
-        io.leak_iop_tokens(2);
-        assert_eq!(io.iop_tokens, 0);
-
-        io.leak_iop_tokens(16000);
-        assert_eq!(io.iop_tokens, 0);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_flush_does_not_consume_iops() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-
-        // Set 0 as IOP limit
-        io.set_iop_limit(16000, 0);
-        assert_none_consumed(&mut io).await;
-
-        let _ = guest
-            .send(BlockOp::Flush {
-                snapshot_details: None,
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Flush {
-                snapshot_details: None,
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Flush {
-                snapshot_details: None,
-            })
-            .await;
-
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-
-        assert_none_consumed(&mut io).await;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_set_bw_limit() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-        io.set_bw_limit(1024 * 1024); // 1 KiB
-
-        assert_none_consumed(&mut io).await;
-
-        // Don't use guest.read, that will send a block size query that will
-        // never be answered.
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(1024, 512), // 512 KiB
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(1024, 512),
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(1024, 512),
-            })
-            .await;
-
-        // First two reads succeed
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-
-        // Next cannot be consumed until there's available BW tokens so it
-        // remains in the queue.
-        assert_none_consumed(&mut io).await;
-        assert!(io.req_rx.try_recv().is_err());
-        assert!(io.req_head.is_some());
-        assert_eq!(io.bw_tokens, 1024 * 1024);
-
-        // Replenish enough tokens, meaning next read can be consumed
-        io.leak_bw_tokens(1024 * 1024 / 2);
-        assert_eq!(io.bw_tokens, 1024 * 1024 / 2);
-
-        assert_consumed(&mut io).await;
-        assert!(io.req_rx.try_recv().is_err());
-        assert!(io.req_head.is_none());
-        assert_eq!(io.bw_tokens, 1024 * 1024);
-
-        io.leak_bw_tokens(1024 * 1024);
-        assert_eq!(io.bw_tokens, 0);
-
-        io.leak_bw_tokens(1024 * 1024 * 1024);
-        assert_eq!(io.bw_tokens, 0);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_flush_does_not_consume_bw() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-
-        // Set 0 as bandwidth limit
-        io.set_bw_limit(0);
-        assert_none_consumed(&mut io).await;
-
-        let _ = guest
-            .send(BlockOp::Flush {
-                snapshot_details: None,
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Flush {
-                snapshot_details: None,
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Flush {
-                snapshot_details: None,
-            })
-            .await;
-
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-        assert_consumed(&mut io).await;
-
-        assert_none_consumed(&mut io).await;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_iop_and_bw_limit() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-
-        io.set_iop_limit(16384, 500); // 1 IOP is 16 KiB
-        io.set_bw_limit(6400 * 1024); // 16384 B * 400 = 6400 KiB/s
-        assert_none_consumed(&mut io).await;
-
-        // Don't use guest.read, that will send a block size query that will
-        // never be answered.
-
-        // Validate that BW limit activates by sending two 7000 KiB IOs. 7000
-        // KiB is only 437.5 IOPs
-
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(14000, 512), // 7000 KiB
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(14000, 512), // 7000 KiB
-            })
-            .await;
-
-        assert_consumed(&mut io).await;
-        assert_none_consumed(&mut io).await;
-
-        // Assert we've hit the BW limit before IOPS
-        assert_eq!(io.iop_tokens, 438); // 437.5 rounded up
-        assert_eq!(io.bw_tokens, 7000 * 1024);
-
-        io.leak_iop_tokens(438);
-        io.leak_bw_tokens(7000 * 1024);
-
-        assert_consumed(&mut io).await;
-
-        // Everything should be empty now
-        assert!(io.req_rx.try_recv().is_err());
-        assert!(io.req_head.is_none());
-
-        // Back to zero
-        io.leak_iop_tokens(438);
-        io.leak_bw_tokens(7000 * 1024);
-
-        assert_eq!(io.iop_tokens, 0);
-        assert_eq!(io.bw_tokens, 0);
-
-        // Validate that IOP limit activates by sending 501 1024b IOs
-        for _ in 0..500 {
-            let _ = guest
-                .send(BlockOp::Read {
-                    offset: Block::new_512(0),
-                    data: Buffer::new(2, 512),
-                })
-                .await;
-            assert_consumed(&mut io).await;
-        }
-
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(2, 512),
-            })
-            .await;
-        assert_none_consumed(&mut io).await;
-
-        // Assert we've hit the IOPS limit
-        assert_eq!(io.iop_tokens, 500);
-        assert_eq!(io.bw_tokens, 500 * 1024);
-
-        // Back to zero
-        io.leak_iop_tokens(500);
-        io.leak_bw_tokens(500 * 1024);
-
-        // Remove the 501st request
-        assert!(io.req_head.take().is_some());
-        assert!(io.req_rx.try_recv().is_err());
-        assert_eq!(io.iop_tokens, 0);
-        assert_eq!(io.bw_tokens, 0);
-
-        // From
-        // https://aws.amazon.com/premiumsupport/knowledge-center/ebs-calculate-optimal-io-size/:
-        //
-        // Amazon EBS calculates the optimal I/O size using the following
-        // equation: throughput / number of IOPS = optimal I/O size.
-
-        let optimal_io_size: usize = 6400 * 1024 / 500;
-
-        // Round down to the nearest size in blocks
-        let optimal_io_size = (optimal_io_size / 512) * 512;
-
-        // Make sure this is <= an IOP size
-        assert!(optimal_io_size <= 16384);
-
-        // I mean, it makes sense: now we submit 500 of those to reach both
-        // limits at the same time.
-        for i in 0..500 {
-            assert_eq!(io.iop_tokens, i);
-            assert_eq!(io.bw_tokens, i * optimal_io_size);
-            assert_eq!(optimal_io_size % 512, 0);
-
-            let _ = guest
-                .send(BlockOp::Read {
-                    offset: Block::new_512(0),
-                    data: Buffer::new(optimal_io_size / 512, 512),
-                })
-                .await;
-
-            assert_consumed(&mut io).await;
-        }
-
-        assert_eq!(io.iop_tokens, 500);
-        assert_eq!(io.bw_tokens, 500 * optimal_io_size);
-
-        Ok(())
-    }
-
-    // Is it possible to submit an IO that will never be sent? It shouldn't be!
-    #[tokio::test]
-    async fn test_impossible_io() -> Result<()> {
-        let (guest, mut io) = Guest::new(None);
-
-        io.set_iop_limit(1024 * 1024 / 2, 10); // 1 IOP is half a KiB
-        io.set_bw_limit(1024 * 1024); // 1 KiB
-        assert_none_consumed(&mut io).await;
-
-        // Sending an IO of 10 MiB is larger than the bandwidth limit and
-        // represents 20 IOPs, larger than the IOP limit.
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(20480, 512), // 10 MiB
-            })
-            .await;
-        let _ = guest
-            .send(BlockOp::Read {
-                offset: Block::new_512(0),
-                data: Buffer::new(0, 512),
-            })
-            .await;
-
-        assert_eq!(io.iop_tokens, 0);
-        assert_eq!(io.bw_tokens, 0);
-
-        // Even though the first IO is larger than the bandwidth and IOP limit,
-        // it should still succeed. The next IO should not, even if it consumes
-        // nothing, because the iops and bw tokens will be larger than the limit
-        // for a while (until they leak enough).
-
-        assert_consumed(&mut io).await;
-        assert_none_consumed(&mut io).await;
-
-        assert_eq!(io.iop_tokens, 20);
-        assert_eq!(io.bw_tokens, 10 * 1024 * 1024);
-
-        // Bandwidth trigger is going to be larger and need more leaking to get
-        // down to a point where the zero sized IO can fire.
-        for _ in 0..9 {
-            io.leak_iop_tokens(10);
-            io.leak_bw_tokens(1024 * 1024);
-
-            assert_none_consumed(&mut io).await;
-        }
-
-        assert_eq!(io.iop_tokens, 0);
-        assert_eq!(io.bw_tokens, 1024 * 1024);
-
-        assert_none_consumed(&mut io).await;
-
-        io.leak_iop_tokens(10);
-        io.leak_bw_tokens(1024 * 1024);
-
-        // We've leaked 10 KiB worth, it should fire now!
-        assert_consumed(&mut io).await;
 
         Ok(())
     }
